--- .vscode/.gitignore
@@ -1,3 +1,0 @@
-/node_modules
-.serverless
-jspm_packages
