--- .vscode/jest.config.js
@@ -1,26 +1,0 @@
-module.exports = {
-    "transform": {
-        ".(ts|tsx)": "ts-jest"
-      },
-      "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
-      "moduleFileExtensions": [
-        "ts",
-        "tsx",
-        "js"
-      ],
-      "coveragePathIgnorePatterns": [
-        "/node_modules/",
-        "/test/"
-      ],
-      "coverageThreshold": {
-        "global": {
-          "branches": 90,
-          "functions": 95,
-          "lines": 95,
-          "statements": 95
-        }
-      },
-      "reporters":[
-        "jest-spec-reporter"
-    ]
-}
