--- .vscode/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "gulp-jsontoxml",
-  "version": "1.0.2",
-  "repository": "https://github.com/gulpetl/gulp-jsontoxml",
-  "description": "Convert xml files to json with gulp",
-  "main": "dist/js/plugin.js",
-  "typings": "dist/types/plugin.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "test": "jest"
-  },
-  "author": "peterl@sbcotton.com",
-  "contributors": [
-    "Manoj Khatri"
-  ],
-  "license": "MIT",
-  "keywords": [
-    "gulpplugin",
-    "gulp-etl",
-    "gulp",
-    "xml",
-    "json",
-    "js",
-    "convert",
-    "transform",
-    "json to xml"
-  ],
-  "homepage": "https://gulpetl.com",
-  "dependencies": {
-    "@types/nodemailer": "^6.2.1",
-    "gulp-xmltojson": "git+https://github.com/Mkr1996/gulp-xmltojson.git",
-    "loglevel": "^1.6.3",
-    "map-stream": "0.0.7",
-    "nodemailer": "^6.3.0",
-    "pkginfo": "^0.4.1",
-    "plugin-error": "^1.0.1",
-    "replace-ext": "^1.0.0",
-    "vinyl": "^2.2.0",
-    "xml-js": "^1.6.11"
-  },
-  "devDependencies": {
-    "@types/jest": "^23.3.14",
-    "@types/node": "^10.14.12",
-    "@types/vinyl": "^2.0.3",
-    "gulp": "^4.0.2",
-    "jest": "^24.8.0",
-    "jest-spec-reporter": "^1.0.6",
-    "ts-jest": "^24.0.2",
-    "ts-node": "^7.0.1",
-    "typescript": "^3.5.3"
-  }
-}
