--- .vscode/tests/sum.test.ts
@@ -1,9 +1,0 @@
-import { sum } from './sum';
-
-test('basic', () => {
-  expect(sum()).toBe(0);
-});
-
-test('basic again', () => {
-  expect(sum(1, 2)).toBe(3);
-});
