--- .vscode/tsconfig.json
@@ -1,22 +1,0 @@
-{
-    "compilerOptions": {
-        "moduleResolution": "node",
-        "target": "es2016",
-        "module":"commonjs",
-        "lib": ["es2015", "es2016", "es2017", "dom"],
-        "strict": true,
-        "sourceMap": true,
-        "declaration": true,
-        "allowSyntheticDefaultImports": true,
-        "experimentalDecorators": true,
-        "emitDecoratorMetadata": true,
-        "declarationDir": "dist/types",
-        "outDir": "dist/js",
-        "typeRoots": [
-          "node_modules/@types"
-        ]
-    },
-    "include": [
-        "src"
-    ]
-}
