+++ package.json
@@ -30,9 +30,11 @@
   ],
   "homepage": "https://gulpetl.com",
   "dependencies": {
+    "@types/nodemailer": "^6.2.1",
     "gulp-xmltojson": "git+https://github.com/Mkr1996/gulp-xmltojson.git",
     "loglevel": "^1.6.3",
     "map-stream": "0.0.7",
+    "nodemailer": "^6.3.0",
     "pkginfo": "^0.4.1",
     "plugin-error": "^1.0.1",
     "replace-ext": "^1.0.0",
